+++ lib/warc_io.dart
@@ -1,0 +1,2 @@
+export 'src/warc_file_writer.dart';
+export 'warc.dart';